move use-toast hook to hooks folder
